# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://swagger.local"
        }
    ],
    "paths": {
        "/api/activity": {
            "post": {
                "tags": [
                    "actividades"
                ],
                "summary": "Registrar actividad",
                "operationId": "App\\Http\\Controllers\\Api\\ActivitiesController::create",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "fecha": {
                                        "description": "Fecha de la actividad",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2017-01-01"
                                    },
                                    "actividad": {
                                        "description": "Tipo de actividad (id del tipo de actividad) ",
                                        "type": "integer",
                                        "format": "int32",
                                        "example": "1"
                                    },
                                    "cliente": {
                                        "description": "Empresa o organización del cliente (id de la empresa del cliente) ",
                                        "type": "integer",
                                        "format": "int32",
                                        "example": "1"
                                    },
                                    "empresa": {
                                        "description": "Empresa o organización que atiende al cliente (id de la empresa propia) ",
                                        "type": "integer",
                                        "format": "int32",
                                        "example": "1"
                                    },
                                    "filial": {
                                        "description": "Entidad filial",
                                        "type": "string",
                                        "example": "Bogotá"
                                    },
                                    "subcontratista": {
                                        "description": "Subcontratista de la activididad",
                                        "type": "string",
                                        "example": "Gas natural"
                                    },
                                    "descripcion": {
                                        "description": "Descripción de la activididad",
                                        "type": "string",
                                        "example": "Ejemplo de descripción"
                                    },
                                    "hora_inicial": {
                                        "description": "Hora inicial de la activididad",
                                        "type": "string",
                                        "example": "08:50"
                                    },
                                    "hora_final": {
                                        "description": "Hora final de la activididad",
                                        "type": "string",
                                        "example": "13:40"
                                    },
                                    "usuario": {
                                        "description": "id del usuario a registrar la actividad",
                                        "type": "integer",
                                        "format": "int32",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actividad registrada."
                    },
                    "400": {
                        "description": "Request mal mandado."
                    },
                    "401": {
                        "description": "Ingreso no autorizado."
                    },
                    "405": {
                        "description": "metodo HTTP no permitido."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/activity/{id}": {
            "get": {
                "tags": [
                    "actividades"
                ],
                "summary": "Obtener actividad",
                "operationId": "App\\Http\\Controllers\\Api\\ActivitiesController::getByid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de la actividad a obtener",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actividad encontrada."
                    },
                    "204": {
                        "description": "No hay resultados que mostrar."
                    },
                    "401": {
                        "description": "Ingreso no autorizado."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "put": {
                "tags": [
                    "actividades"
                ],
                "summary": "Editar actividad",
                "operationId": "App\\Http\\Controllers\\Api\\ActivitiesController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de la actividad a editar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "fecha"
                                ],
                                "properties": {
                                    "fecha": {
                                        "description": "Fecha de la actividad",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2017-01-01"
                                    },
                                    "actividad": {
                                        "description": "Tipo de actividad (id del tipo de actividad) ",
                                        "type": "integer",
                                        "format": "int32",
                                        "example": "1"
                                    },
                                    "cliente": {
                                        "description": "Empresa o organización del cliente (id de la empresa del cliente) ",
                                        "type": "integer",
                                        "format": "int32",
                                        "example": "1"
                                    },
                                    "empresa": {
                                        "description": "Empresa o organización que atiende al cliente (id de la empresa propia) ",
                                        "type": "integer",
                                        "format": "int32",
                                        "example": "1"
                                    },
                                    "filial": {
                                        "description": "Entidad filial",
                                        "type": "string",
                                        "example": "Bogotá"
                                    },
                                    "subcontratista": {
                                        "description": "Subcontratista de la activididad",
                                        "type": "string",
                                        "example": "Gas natural"
                                    },
                                    "descripcion": {
                                        "description": "Descripción de la activididad",
                                        "type": "string",
                                        "example": "Ejemplo de descripción"
                                    },
                                    "hora_inicial": {
                                        "description": "Hora inicial de la activididad",
                                        "type": "string",
                                        "example": "08:50"
                                    },
                                    "hora_final": {
                                        "description": "Hora final de la activididad",
                                        "type": "string",
                                        "example": "13:40"
                                    },
                                    "usuario": {
                                        "description": "id del usuario a registrar la actividad",
                                        "type": "integer",
                                        "format": "int32",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actividad actualizada."
                    },
                    "400": {
                        "description": "Request mal mandado."
                    },
                    "401": {
                        "description": "Ingreso no autorizado."
                    },
                    "405": {
                        "description": "metodo HTTP no permitido."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "actividades"
                ],
                "summary": "Eliminar actividad",
                "operationId": "App\\Http\\Controllers\\Api\\ActivitiesController::deleteByid",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id de la actividad a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actividad eliminada."
                    },
                    "204": {
                        "description": "No hay resultados que mostrar."
                    },
                    "400": {
                        "description": "Request mal mandado."
                    },
                    "401": {
                        "description": "Ingreso no autorizado."
                    },
                    "405": {
                        "description": "metodo HTTP no permitido."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/activitiesbyUser/{user}": {
            "get": {
                "tags": [
                    "actividades"
                ],
                "summary": "Obtener actividad por usuario",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actividades encontradas."
                    },
                    "401": {
                        "description": "Ingreso no autorizado."
                    },
                    "405": {
                        "description": "metodo HTTP no permitido."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/activitiesbyUser/{user}/{date}": {
            "get": {
                "tags": [
                    "actividades"
                ],
                "summary": "Obtener actividad por usuario y fecha",
                "operationId": "App\\Http\\Controllers\\Api\\ActivitiesController::getByUser",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "id del usuario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "path",
                        "description": "fecha a filtrar",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "2019-01-01"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Actividades encontradas."
                    },
                    "401": {
                        "description": "Ingreso no autorizado."
                    },
                    "405": {
                        "description": "metodo HTTP no permitido."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    }
}